Added: A Gameobject class that represents our game entities (Ship, Bullet, etc). It is handled in the Game object. Generally, you'll create an instance of your GameObject and call addGameObject on the Game to add it to the logic.
Added: A specific game class for our game that demonstrates the inheritance of game.

Added: A Ship class for demonstrating the use of inheriting from the GameObject class and suggested use of the Sprite class.

Updated: The Sprite management is now functional.
Updated: The game now has a logic and draw phase to it's loop. Drawing operates on the sprite manager, which in turn draws sprites it contains. And the logic phase polls events and notifies it's registered game objects, as well as the super class that inherits the Game. Also, every game object has it's update method called.

Updated: Fixed all usage of creating objects, now everything uses "new Object();" if it's a pointer. The old style created memory initialization problems.

Updated: Simplified the main call. Properties of the Game are passed in through the constructor, thus the super class of Game provides the window caption and window size.

Added: A "Types" file which is meant to contain shortcut type defs for example uint instead of unsigned int.